Some refactoring to tidy up code minimally
