+++ pling/utils.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+
+def get_pling_root_dir() -> Path:
+    return Path(__file__).parent.parent
